+++ MakeMeUpzz/Repository/MakeupTypeRepository.cs
@@ -1,5 +1,6 @@
 ﻿using MakeMeUpzz.Model;
 using MakeMeUpzz.Repository;
+using System;
 using System.Collections.Generic;
 using System.Data.Entity;
 using System.Linq;
@@ -22,31 +23,45 @@ namespace MakeMeUpzz.Repositories
 
         public void AddMakeupType(MakeupType makeupType)
         {
-            makeupType.MakeupTypeID = GenerateMakeupTypeId();
             db.MakeupTypes.Add(makeupType);
             db.SaveChanges();
         }
 
-        public void UpdateMakeupType(MakeupType makeupType)
+        public void UpdateMakeupType(int typeId, string typeName)
         {
-            db.Entry(makeupType).State = EntityState.Modified;
+            MakeupType makeupType = db.MakeupTypes.Find(typeId);
+            makeupType.MakeupTypeName = typeName;
+
             db.SaveChanges();
         }
 
-        public void DeleteMakeupType(int id)
+        public string DeleteMakeupType(int id)
         {
-            var makeupType = db.MakeupTypes.Find(id);
-            if (makeupType != null)
+            MakeupType makeupType = db.MakeupTypes.Find(id);
+            String response = "";
+            bool isUsed = db.Makeups.Any(x => x.MakeupTypeID == id);
+            if (makeupType != null && !isUsed)
             {
                 db.MakeupTypes.Remove(makeupType);
                 db.SaveChanges();
+                response = "Makeup type has been deleted!";
             }
+            else
+            {
+                response = "Please check the used type first!";
+            }
+            return response;
         }
 
-        private int GenerateMakeupTypeId()
+        public int GenerateMakeupTypeId()
         {
             int lastId = db.MakeupTypes.OrderByDescending(x => x.MakeupTypeID).Select(x => x.MakeupTypeID).FirstOrDefault();
             return lastId + 1;
+        }
+
+        public MakeupType GetMakeupTypeByName(string name)
+        {
+            return db.MakeupTypes.Where(x => x.MakeupTypeName == name).FirstOrDefault();
         }
     }
 }
